Fix, bug writing ExecOptions::stdinData, causing it to be frequently lost.
When writing to a byte pipe you close the pipe to signal the consumer that
there's no more data. But when you've written to an async pipe you must wait
until the confirmation of the write before you call close. Otherwise the
close might happen first internally, and the data gets dropped because it's
being written to a closed pipe. When writing to a byte pipe you close the pipe to
signal the consumer that there's no more data. But when you've written to an
async pipe you must wait until the confirmation of the write before you call
close. Otherwise the close might happen first internally, and the data gets
dropped because it's being written to a closed pipe.